10.27更新，版本3.4.4 1.优化了部分未翻译的文字、配置文件 2.添加了主配置文件的英文注释 3.修复了自动更新和检查更新的问题

+++ UltiTools/src/main/java/com/minecraft/ultikits/checker/updatechecker/ConfigFileChecker.java
@@ -1,9 +1,7 @@
 package com.minecraft.ultikits.checker.updatechecker;
 
 import com.minecraft.ultikits.config.ConfigController;
-import com.minecraft.ultikits.enums.ConfigsEnum;
 import com.minecraft.ultikits.ultitools.UltiTools;
-import com.minecraft.ultikits.utils.LanguageUtils;
 import org.bukkit.ChatColor;
 import org.bukkit.configuration.file.YamlConfiguration;
 import org.bukkit.scheduler.BukkitRunnable;
@@ -17,10 +15,9 @@ import java.nio.channels.ReadableByteChannel;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.function.IntFunction;
+import java.util.Objects;
 
-import static com.minecraft.ultikits.checker.updatechecker.VersionChecker.isOutDate;
-import static com.minecraft.ultikits.checker.updatechecker.VersionChecker.version;
+import static com.minecraft.ultikits.checker.updatechecker.VersionChecker.*;
 import static com.minecraft.ultikits.utils.Utils.getFiles;
 
 public class ConfigFileChecker {
@@ -55,7 +52,7 @@ public class ConfigFileChecker {
         File file = new File(UltiTools.getInstance().getDataFolder(), "config.yml");
         transferOldConfigs(config);
         if (file.delete()) {
-            UltiTools.getInstance().saveDefaultConfig();
+            UltiTools.yaml.saveYamlFile(UltiTools.getInstance().getDataFolder().getPath(), "config.yml", UltiTools.language+"_config.yml");
             File newFile = new File(UltiTools.getInstance().getDataFolder(), "config.yml");
             YamlConfiguration newConfig = YamlConfiguration.loadConfiguration(file);
             for (String key : config.keySet()) {
@@ -69,6 +66,15 @@ public class ConfigFileChecker {
                 e.printStackTrace();
             }
         }
+//        boolean change = false;
+//        Configuration defaults = YamlConfiguration.loadConfiguration(new File());
+//        for (String defaultKey : defaults.getKeys(true)) {
+//            if (!UltiTools.getInstance().getConfig().contains(defaultKey)) {
+//                UltiTools.getInstance().getConfig().set(defaultKey, defaults.get(defaultKey));
+//                change = true;
+//            }
+//        }
+//        if (change) UltiTools.getInstance().saveConfig();
     }
 
     public static void transferOldConfigs(Map<String, Object> oldConfig){
@@ -121,10 +127,12 @@ public class ConfigFileChecker {
         }
 
         public static void deleteOldVersion () {
-            List<File> files = getFiles(UltiTools.getInstance().getDataFolder().getPath().replace(java.io.File.separator+"UltiTools", ""));
-            for (File file : files) {
-                if (file.getName().contains("UltiTools-") && !file.getName().equals("UltiTools-" + version + ".jar")) {
-                    file.delete();
+            if (!Objects.equals(current_version, version)) {
+                List<File> files = getFiles(UltiTools.getInstance().getDataFolder().getPath().replace(java.io.File.separator + "UltiTools", ""));
+                for (File file : files) {
+                    if (file.getName().contains("UltiTools-") && !file.getName().equals("UltiTools-" + version + ".jar")) {
+                        file.delete();
+                    }
                 }
             }
         }
